Move Category and MonoidalCategory to discopy/abc.py
The abstract base classes have no discopy dependencies, so giving them
their own low-level module lets every module import them without import
cycles. In particular Drawing and Hypergraph, which sit below monoidal
but need the @ tensor sugar, can subclass MonoidalCategory from
discopy.abc instead of pulling in monoidal.

cat.py, monoidal.py and the others now import Category/MonoidalCategory
from discopy.abc.
