+++ discopy/stream.py
@@ -166,7 +166,7 @@ from typing import Callable, Optional
 from dataclasses import dataclass
 
 from discopy import symmetric
-from discopy.cat import MonoidalCategory
+from discopy.abc import MonoidalCategory
 from discopy.utils import (
     AxiomError, NamedGeneric, get_origin, is_tuple,
     assert_isinstance, unbiased, inductive, classproperty, factory_name)
